requirements.txt
apache-airflow
requests
pandas
pyarrow
fastparquet